discogs_credits: extract logger and resolveLinkTypeId
Issue #32 — two related extractions:

  src/log.js
    addLogLine, setLogContainer, getLogContainer.
    Container is wired lazily by ui-bar code at insertion time (the <ul>
    is created inside the bar's mount routine); pre-init calls silently
    no-op rather than throwing.

  src/api-mb.js (extended)
    resolveLinkTypeId moved here — depends on `addLogLine`, which is now
    importable. Imports pageWindow from constants.js.

Entry file no longer contains either definition; both are imported.
Full 12-fixture suite green; no behavior change.
