+++ userscripts/discogs_credits/dev/DECISIONS.md
@@ -79,4 +79,5 @@ For the full reasoning behind any decision, `git log -- dev/ANALYSIS.md` finds t
 - 2026-05-24 22:12 — issue #32 chunk 2b (constants + util) → extracted `REL_TEMPLATE`, `SELECTORS`, `DISCOGS_LOGO_URL` to `src/constants.js`; the five zero-dep DOM helpers (`doNext`, `setNativeValue`, `selectValue`, `makeKeyDownEvent`, `makeClickEvent`) to `src/util.js`. ~100 lines moved out. Entry file now down to 3713 lines (from 4983 originally; the original headroom was data + UI sugar). Smoke fixture passes; full suite running.
 - 2026-05-24 22:17 — issue #32 chunk 2c (Discogs API + URL helpers) → extracted `link_infos`, `getDiscogsLinkKey`, `_releaseDataCache`, and `getDiscogsReleaseData` to `src/api-discogs.js`. `link_infos` stays exported because many consumers in the entry still read `link_infos[key]` directly (this will move to a parser-returns-object shape per ANALYSIS §2.3 once the consumers themselves are moved). Smoke fixture passes; full suite running.
 - 2026-05-24 22:43 — issue #32 chunk 2d (MB throttle + page-window const) → extracted `mbThrottle`, `fetchMBEntity`, and `fetchWithRetry` to `src/api-mb.js`. Moved `pageWindow` from the entry top into `constants.js` so api-mb (and future modules) can share the single derived reference without re-evaluating `typeof unsafeWindow`. `resolveLinkTypeId` stays in entry for now — it calls `addLogLine`, which lives in the not-yet-extracted logger. Will move once the logger is its own module. Smoke fixture passes; full suite running.
+- 2026-05-24 22:50 — issue #32 chunk 2e (logger + resolveLinkTypeId) → extracted `addLogLine` to `src/log.js`. The log container (the `<ul>` element the UI bar creates at startup) is now wired in via `setLogContainer(el)` after the UI bar constructs it; this lets log.js be a standalone module without circular access to the entry's `let logs`. Pre-init `addLogLine` calls silently no-op instead of throwing (slightly safer than today). With the logger out, `resolveLinkTypeId` follows it into `src/api-mb.js` (it depends on `addLogLine`). Entry no longer has either function definition. Smoke fixture passes; full suite running.
 - 2026-05-24 19:55 — issue #30 (entity-fetch slower than the old "5-thread burst" approach) → `mbThrottle` rewritten. The previous shape was a strict serial promise chain with a post-success `setTimeout(_gap)` between each request (`_gap` ∈ [200, 3000] ms), so the 5-worker pools in `checkMissingArtists` / `checkMissingCompanies` all funnelled through one queue at ≤5 req/s — slower than the old "5 immediate parallel threads" approach the throttle was meant to replace. New shape: up to `MAX_CONCURRENT=4` in-flight requests, no artificial gap, **global cooperative `_pauseUntil`** on 429/503 that honors `Retry-After` (or exponential per-request backoff if absent). All in-flight workers idle together until the pause elapses, which avoids the thundering-herd retries that the original 5-thread approach produced. Also removed two redundant `setTimeout(200)` calls (showReviewTable name pre-load loop and the URL-check pool) — the throttle paces. Expected ~3–5× speedup on rate-limited-heavy fixtures (Midwest Funk-style preflight), without 503 storms.
